backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
email-validator
bcrypt==3.2.2
